+++ app/Modules/Projects/Enums/ProjectType.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Modules\Projects\Enums;
+
+enum ProjectType: string
+{
+    case Fixed = 'fixed';
+    case Hourly = 'hourly';
+    case Retainer = 'retainer';
+}